# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HouseholdCreate": {
        "properties": {
          "locale": {
            "default": "es",
            "maxLength": 32,
            "minLength": 2,
            "title": "Locale",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "HouseholdCreate",
        "type": "object"
      },
      "HouseholdResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "locale",
          "timezone",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "HouseholdResponse",
        "type": "object"
      },
      "HouseholdUpdate": {
        "properties": {
          "locale": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "HouseholdUpdate",
        "type": "object"
      },
      "HouseholdWithMembership": {
        "properties": {
          "household": {
            "$ref": "#/components/schemas/HouseholdResponse"
          },
          "membership": {
            "$ref": "#/components/schemas/MembershipResponse"
          }
        },
        "required": [
          "household",
          "membership"
        ],
        "title": "HouseholdWithMembership",
        "type": "object"
      },
...
      "MemberPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page_info"
        ],
        "title": "MemberPage",
        "type": "object"
      },
      "MemberResponse": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "joined_at": {
            "format": "date-time",
            "title": "Joined At",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "display_name",
          "email",
          "email_verified",
          "role",
          "status",
          "version",
          "joined_at"
        ],
        "title": "MemberResponse",
        "type": "object"
      },
      "MemberRoleUpdate": {
        "properties": {
          "role": {
            "enum": [
              "admin",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "MemberRoleUpdate",
        "type": "object"
      },
      "MembershipResponse": {
        "properties": {
          "household_id": {
            "format": "uuid",
            "title": "Household Id",
            "type": "string"
          },
          "household_name": {
            "title": "Household Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "joined_at": {
            "format": "date-time",
            "title": "Joined At",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "household_id",
          "household_name",
          "role",
          "status",
          "version",
          "joined_at"
        ],
        "title": "MembershipResponse",
        "type": "object"
      },
      "PageInfo": {
        "properties": {
          "limit": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "PageInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Uribap household food planning API",
    "title": "Uribap API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/households": {
      "post": {
        "operationId": "create_api_v1_households_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdWithMembership"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{household_id}": {
      "get": {
        "operationId": "read_api_v1_households__household_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read",
        "tags": [
          "households"
        ]
      },
      "patch": {
        "operationId": "update_api_v1_households__household_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update",
        "tags": [
          "households"
        ]
      }
    },
...
    "/api/v1/households/{household_id}/members": {
      "get": {
        "operationId": "members_api_v1_households__household_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Members",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{household_id}/members/{user_id}": {
      "delete": {
        "operationId": "delete_member_api_v1_households__household_id__members__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Member",
        "tags": [
          "households"
        ]
      },
      "patch": {
        "operationId": "update_member_api_v1_households__household_id__members__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Member",
        "tags": [
          "households"
        ]
      }
    },
...
  }
}